fix test_pipeline_fit_time_constraint (#633)
Fixed occasional test error when computational time of pipeline with higher time constraint was lower than computational time with less time constraint
Time constraint affects fit time, but fit time also fluctuates because of internal hardware reasons. The dispersion of this fluctuation seems to be constant. Therefore, if fit process is optimized, the fit time of both time constraint values in test is reduced, but dispersion remains the same. This raises the probability of dispersions' overlay and, hence, the test fail. Change of timeouts to 1 and 5 seconds provides stable work of test